fixed: Replaced cart.ts to domains, used Map instead of Array, fixed logic of addProducts and renamed methods to correct names
